fix(backend): rename register DTO file
