+++ backend/src/auth/dto/register.dto.ts
