+++ backend/src/certificates/entities/certificate.entity.ts
@@ -26,8 +26,8 @@ export class Certificate {
   /**
    * Recipient info (denormalized for easy access & PDF rendering)
    */
-  @Column()
-  recipientName: string;
+  @Column({ nullable: true })
+  recipientName: string | null;
 
   @Column()
   recipientEmail: string;
