+++ backend/src/users/users.service.ts
@@ -9,7 +9,7 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { Repository, MoreThan } from 'typeorm';
 import * as crypto from 'crypto';
 import * as bcrypt from 'bcrypt';
-import { RegisterDto } from 'src/auth/dto/ register.dto';
+import { RegisterDto } from 'src/auth/dto/register.dto';
 import { UpdateProfileDto } from './dto/update-profile.dto';
 import { User, UserRole } from './entities/user.entity';
 import { Certificate } from '../certificates/entities/certificate.entity';
